mgr/cephadm: move _refresh_host_devices to serve.py
